Refactor handling of letter not found
