+++ internal/datastore/src/errors/letter-already-exists-error.ts
